Basic Multiversion Fabric support
