+++ fabric/mc1201/src/main/resources/fabric.mod.json
@@ -1,0 +1,35 @@
+{
+  "schemaVersion": 1,
+  "id": "packetevents-mc1201",
+  "version": "${version}",
+  "name": "PacketEvents for 1.20.1",
+  "description": "",
+  "authors": [
+    "retrooper"
+  ],
+  "contact": {},
+  "license": "GPL-3.0",
+  "environment": "*",
+  "entrypoints": {
+    "pePreLaunchClient": [
+      "io.github.retrooper.packetevents.mc1201.PacketEventsClientMod"
+    ],
+    "pePreLaunchServer": [
+      "io.github.retrooper.packetevents.mc1201.PacketEventsServerMod"
+    ],
+    "preLaunch": [
+      "io.github.retrooper.packetevents.mc1201.PacketEventsMod"
+    ],
+    "main": [
+      "io.github.retrooper.packetevents.mc1201.PacketEventsMod"
+    ]
+  },
+  "mixins": [
+    "packetevents-mc1201.mixins.json"
+  ],
+  "accessWidener": "packetevents.accesswidener",
+  "depends": {
+    "fabricloader": ">=0.15.11",
+    "minecraft": ">=1.20 <1.21"
+  }
+}
