+++ fabric/mc1211/gradle.properties
@@ -4,7 +4,7 @@ org.gradle.parallel=true
 
 # Fabric Properties
 # check these on https://fabricmc.net/develop
-minecraft_version=1.20.1
-parchment_mappings=2023.09.03
-parchment_minecraft_version=1.20.1
+minecraft_version=1.21.1
+parchment_mappings=2024.11.17
+parchment_minecraft_version=1.21.1
 loader_version=0.15.11
