+++ fabric/mc1211/src/client/java/io/github/retrooper/packetevents/mc1211/PacketEventsClientMod.java
@@ -16,26 +16,18 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package io.github.retrooper.packetevents;
+package io.github.retrooper.packetevents.mc1211;
 
-import com.github.retrooper.packetevents.PacketEvents;
+
+import io.github.retrooper.packetevents.LazyHolder;
 import io.github.retrooper.packetevents.factory.fabric.FabricPacketEventsAPI;
-import io.github.retrooper.packetevents.impl.netty.manager.player.PlayerManagerAbstract;
-import java.util.function.Supplier;
-import net.fabricmc.api.EnvType;
-import net.fabricmc.loader.api.entrypoint.PreLaunchEntrypoint;
+import io.github.retrooper.packetevents.mc1211.factory.fabric.Fabric1211ClientPlayerManager;
 
-public class PacketEventsServerMod implements PreLaunchEntrypoint {
-
-    public static FabricPacketEventsAPI constructApi(String modid) {
-        return new FabricPacketEventsAPI(modid, EnvType.SERVER);
-    }
+public class PacketEventsClientMod extends io.github.retrooper.packetevents.PacketEventsClientMod {
 
     @Override
-    public void onPreLaunch() {}
-
-    public void preLaunch() {
-        PacketEvents.setAPI(constructApi(PacketEventsMod.MOD_ID));
-        PacketEvents.getAPI().load();
+    public void onPreLaunch() {
+        FabricPacketEventsAPI.staticLazyPlayerManagerHolder = LazyHolder.simple(Fabric1211ClientPlayerManager::new);
+        super.preLaunch();
     }
 }
