+++ fabric/mc1211/src/main/java/io/github/retrooper/packetevents/mc1211/PacketEventsMod.java
@@ -16,26 +16,17 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package io.github.retrooper.packetevents;
+package io.github.retrooper.packetevents.mc1211;
 
-import com.github.retrooper.packetevents.PacketEvents;
-import io.github.retrooper.packetevents.factory.fabric.FabricPacketEventsAPI;
-import io.github.retrooper.packetevents.impl.netty.manager.player.PlayerManagerAbstract;
-import java.util.function.Supplier;
-import net.fabricmc.api.EnvType;
-import net.fabricmc.loader.api.entrypoint.PreLaunchEntrypoint;
+public class PacketEventsMod extends io.github.retrooper.packetevents.PacketEventsMod {
 
-public class PacketEventsServerMod implements PreLaunchEntrypoint {
-
-    public static FabricPacketEventsAPI constructApi(String modid) {
-        return new FabricPacketEventsAPI(modid, EnvType.SERVER);
+    @Override
+    public void onPreLaunch() {
+        super.preLaunch();
     }
 
     @Override
-    public void onPreLaunch() {}
-
-    public void preLaunch() {
-        PacketEvents.setAPI(constructApi(PacketEventsMod.MOD_ID));
-        PacketEvents.getAPI().load();
+    public void onInitialize() {
+        super.initialize();
     }
 }
