+++ fabric/mc1211/src/main/java/io/github/retrooper/packetevents/mc1211/factory/fabric/Fabric1211ServerPlayerManager.java
@@ -16,26 +16,27 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package io.github.retrooper.packetevents;
+package io.github.retrooper.packetevents.mc1211.factory.fabric;
 
-import com.github.retrooper.packetevents.PacketEvents;
-import io.github.retrooper.packetevents.factory.fabric.FabricPacketEventsAPI;
 import io.github.retrooper.packetevents.impl.netty.manager.player.PlayerManagerAbstract;
-import java.util.function.Supplier;
-import net.fabricmc.api.EnvType;
-import net.fabricmc.loader.api.entrypoint.PreLaunchEntrypoint;
+import net.minecraft.server.level.ServerPlayer;
+import org.jetbrains.annotations.NotNull;
 
-public class PacketEventsServerMod implements PreLaunchEntrypoint {
+public class Fabric1211ServerPlayerManager extends PlayerManagerAbstract {
 
-    public static FabricPacketEventsAPI constructApi(String modid) {
-        return new FabricPacketEventsAPI(modid, EnvType.SERVER);
+    @Override
+    public int getPing(@NotNull Object player) {
+        if (player instanceof ServerPlayer) {
+            return ((ServerPlayer) player).connection.latency();
+        }
+        throw new UnsupportedOperationException("Unsupported player implementation: " + player);
     }
 
     @Override
-    public void onPreLaunch() {}
-
-    public void preLaunch() {
-        PacketEvents.setAPI(constructApi(PacketEventsMod.MOD_ID));
-        PacketEvents.getAPI().load();
+    public Object getChannel(@NotNull Object player) {
+        if (player instanceof ServerPlayer) {
+            return ((ServerPlayer) player).connection.connection.channel;
+        }
+        throw new UnsupportedOperationException("Unsupported player implementation: " + player);
     }
 }
