+++ fabric/mc1211/src/main/resources/packetevents-mc1211.mixins.json
@@ -1,0 +1,16 @@
+{
+  "required": true,
+  "minVersion": "0.8",
+  "package": "io.github.retrooper.packetevents.mc1201.mixin",
+  "compatibilityLevel": "JAVA_17",
+  "client": [
+    "ClientPacketListenerMixin"
+  ],
+  "mixins": [
+    "ConnectionMixin",
+    "PlayerListMixin"
+  ],
+  "injectors": {
+    "defaultRequire": 1
+  }
+}
