+++ fabric/mc1211/src/main/resources/packetevents.accesswidener
@@ -1,0 +1,11 @@
+accessWidener v1 named
+
+
+# fields
+accessible field net/minecraft/network/Connection channel Lio/netty/channel/Channel;
+accessible field net/minecraft/server/network/ServerCommonPacketListenerImpl connection Lnet/minecraft/network/Connection;
+accessible field net/minecraft/client/multiplayer/ClientCommonPacketListenerImpl connection Lnet/minecraft/network/Connection;
+
+# methods
+
+# classes
