+++ fabric/src/main/java/io/github/retrooper/packetevents/LazyHolder.java
@@ -1,0 +1,57 @@
+package io.github.retrooper.packetevents;
+
+import java.util.function.Supplier;
+
+public interface LazyHolder<T> {
+    static <T> LazyHolder<T> threadSafe(Supplier<T> supplier) {
+        return new ThreadSafeLazyHolder<>(supplier);
+    }
+
+    static <T> LazyHolder<T> simple(Supplier<T> supplier) {
+        return new SimpleLazyHolder<>(supplier);
+    }
+
+    T get();
+}
+
+final class ThreadSafeLazyHolder<T> implements LazyHolder<T> {
+    private final Supplier<T> supplier;
+    private volatile T value;
+
+    ThreadSafeLazyHolder(Supplier<T> supplier) {
+        this.supplier = supplier;
+    }
+
+    @Override
+    public T get() {
+        T result = value;
+        if (result == null) {
+            synchronized (this) {
+                result = value;
+                if (result == null) {
+                    result = supplier.get();
+                    value = result;
+                }
+            }
+        }
+        return result;
+    }
+}
+
+final class SimpleLazyHolder<T> implements LazyHolder<T> {
+    private T value;
+    private Supplier<T> supplier;
+
+    SimpleLazyHolder(Supplier<T> supplier) {
+        this.supplier = supplier;
+    }
+
+    @Override
+    public T get() {
+        if (supplier != null) {
+            value = supplier.get();
+            supplier = null;
+        }
+        return value;
+    }
+}
