+++ fabric/src/main/java/io/github/retrooper/packetevents/factory/fabric/FabricPacketEventsAPI.java
@@ -29,6 +29,7 @@ import com.github.retrooper.packetevents.netty.NettyManager;
 import com.github.retrooper.packetevents.protocol.packettype.PacketType;
 import com.github.retrooper.packetevents.settings.PacketEventsSettings;
 import com.github.retrooper.packetevents.util.LogManager;
+import io.github.retrooper.packetevents.LazyHolder;
 import io.github.retrooper.packetevents.impl.netty.NettyManagerImpl;
 import io.github.retrooper.packetevents.impl.netty.manager.player.PlayerManagerAbstract;
 import net.fabricmc.api.EnvType;
@@ -41,6 +42,9 @@ import java.util.Locale;
 public class FabricPacketEventsAPI extends PacketEventsAPI<FabricLoader> {
 
     private static final Logger LOGGER = LoggerFactory.getLogger("PacketEvents");
+    // TODO, refactor if booky and retrooper approve, bad design having settable static field, exists to maintain
+    // 100% backward compatability
+    public static LazyHolder<PlayerManagerAbstract> staticLazyPlayerManagerHolder = () -> null;
 
     private final String modId;
     private final EnvType environment;
@@ -48,7 +52,7 @@ public class FabricPacketEventsAPI extends PacketEventsAPI<FabricLoader> {
 
     private final ProtocolManager protocolManager;
     private final ServerManager serverManager;
-    private final PlayerManagerAbstract playerManager;
+    private final LazyHolder<PlayerManagerAbstract> playerManager;
     private final ChannelInjector injector;
     private final NettyManager nettyManager = new NettyManagerImpl();
     private final LogManager logManager = new FabricLogger(LOGGER);
@@ -68,16 +72,12 @@ public class FabricPacketEventsAPI extends PacketEventsAPI<FabricLoader> {
 
         this.protocolManager = new FabricProtocolManager(environment);
         this.serverManager = this.constructServerManager();
-        this.playerManager = this.constructPlayerManager();
+        this.playerManager = null;
         this.injector = new FabricChannelInjector(environment);
     }
 
     protected ServerManager constructServerManager() {
         return new FabricServerManager();
-    }
-
-    protected PlayerManagerAbstract constructPlayerManager() {
-        return new FabricPlayerManager();
     }
 
     @Override
@@ -164,7 +164,7 @@ public class FabricPacketEventsAPI extends PacketEventsAPI<FabricLoader> {
 
     @Override
     public PlayerManager getPlayerManager() {
-        return this.playerManager;
+        return staticLazyPlayerManagerHolder.get();
     }
 
     @Override
