+++ fabric/src/main/resources/packetevents.accesswidener
@@ -2,9 +2,8 @@ accessWidener v1 named
 
 
 # fields
-accessible field net/minecraft/server/network/ServerCommonPacketListenerImpl connection Lnet/minecraft/network/Connection;
 accessible field net/minecraft/network/Connection channel Lio/netty/channel/Channel;
-
+accessible field net/minecraft/server/network/ServerGamePacketListenerImpl connection Lnet/minecraft/network/Connection;
 # methods
 
 # classes
